fixed null ref bugs

+++ Models/LinuxServiceMetrics.cs
@@ -6,7 +6,7 @@ namespace zoneswitch.metricsgenerator.Models
         public string ServiceName { get; set; }
         public long RequestRate { get; set; }
         public long ResponseRate { get; set; }
-        public int SuccessRate { get; set; }
+        public double SuccessRate { get; set; }
         public string FunctionName { get; set; }
     }
 }
